refactor : PrePersist 기능을 제거하고 Auditing 기능으로 변경 및 공통 필드 매핑에 대한 책임 수정
