+++ src/main/java/com/prgms/jpaBoard/JpaBoardApplication.java
@@ -2,8 +2,10 @@ package com.prgms.jpaBoard;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
+@EnableJpaAuditing
 public class JpaBoardApplication {
 
 	public static void main(String[] args) {
