+++ src/main/java/com/prgms/jpaBoard/global/createdUser.java
@@ -1,0 +1,14 @@
+package com.prgms.jpaBoard.global;
+
+public class createdUser extends createdTime {
+
+    private String createdBy;
+
+    public void setCreatedBy(String createdBy) {
+        this.createdBy = createdBy;
+    }
+
+    public String getCreatedBy() {
+        return createdBy;
+    }
+}
